% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_missingness.R
\name{add_missingness}
\alias{add_missingness}
\title{add_missingness function}
\usage{
add_missingness(G, missing_rate, seed = NA)
}
\arguments{
\item{G}{: A binary, numeric genome matrix.}

\item{missing_rate}{: Probability for any given element of G to become NA.}

\item{seed}{: Optional. A random seed for reproducibility.}
}
\value{
A genome matrix with NA elements.
}
\description{
Randomly adds missingness to a genome matrix. Each element in the matrix becomes NA with
a fixed probability (missing_rate). The trials are independent for each matrix element.
}
\examples{
  G = matrix(sample(0:1, size = 25, replace = TRUE), nc = 5)
add_missingness(G)
}
