% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/age_DNA.R
\name{age_DNA}
\alias{age_DNA}
\title{age_DNA function}
\usage{
age_DNA(G, missing_rate, trans_rate = 0.776, dmg_rate = 0.05, seed = NA)
}
\arguments{
\item{G}{: A binary, numeric genome matrix from discoal.}

\item{missing_rate}{: Probability of elements that are randomly sampled to become NA for each row in G.}

\item{trans_rate:}{Probability of elements that are randomly sampled to be transitions for each row in G. 
Default is 0.776.}

\item{dmg_rate:}{Probability that a randomly sampled transition with an ancestral allele 0 is convert to 1.
Default is 0.05.}

\item{seed:}{A numeric random seed. Optional.}
}
\value{

}
\description{
A function to simulating DNA aging on discoal simulations. To simulate missingness, for each row 
a percentage of elements are randomly sampled and become missing. To simulate deamination
a percentage of elements in each row are randomly sampled according to trans_rate. Then 
if the element is 0, it becomes 1 with probability designated by  dmg_rate.
}
\details{
doi: 10.1534/genetics.112.139949
}
\examples{

}
