% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ancient_generate_df.R
\name{ancient_generate_df}
\alias{ancient_generate_df}
\title{ancient_generate_df function}
\usage{
ancient_generate_df(
  sim_list,
  nwins,
  split_type = "base",
  trim_sim = T,
  snp = NA,
  missing_rate,
  trans_prop = 0.776,
  dmg_rate = 0.05,
  index,
  seed = NA
)
}
\arguments{
\item{snp}{: Optional. Number of snps to include per simulation. Used only when trim_sim is T.}

\item{missing_rate}{: Probability of elements that are randomly sampled to become NA for each row in G.}

\item{sim_list:}{list of simulation objects}

\item{nwins:}{number of subwindows desired to split each genome matrix per simulation}

\item{split_type:}{Method of splitting the genome matrix. Valid options are "base" and "mut". Default "base".}

\item{trim_sim:}{a logical indicating whether outer snps should be removed. Default is F.}

\item{trans_prop:}{Proportion of columns (i.e. sites) that are chosen to be transition sites.
Default is 0.776.}

\item{dmg_rate:}{Probability of a element in a transition column changing from 0 to 1, or 1 to 0. 
Default is 0.05.}

\item{index:}{A vector of indices indicating the rows of the genome matrix which represent the outgroup.}

\item{seed:}{Optional. A random seed for aging DNA.}
}
\value{
a dataframe containting the summary statistics for the list of simulation objects
}
\description{
Takes in a list of simulation objects, simulates DNA aging and generates
summary statistics. See the generate_df documentation for more information.
}
\examples{
generate_df(sim_list,nwins = 10, missing_rate = 0.05, index = c(99,100))
}
