% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ancient_sum_stats.R
\name{ancient_sum_stats}
\alias{ancient_sum_stats}
\title{ancient_sum_stats function}
\usage{
ancient_sum_stats(
  sim,
  nwins = 1,
  split_type = "base",
  ID,
  trim_sim = T,
  snp = NA,
  form = "wide",
  LD_downsample = F,
  ds_prop = NA,
  ds_seed = NA
)
}
\arguments{
\item{sim:}{a simulation object}

\item{nwins:}{number of subwindows to split each genome matrix within the simulation. Default is 1.}

\item{split_type:}{Method of splitting the genome matrix. Valid options are "base" and "mut". Default is "base".}

\item{ID:}{a numeric ID value to group subwindows under the simulation it came from.}

\item{trim_sim:}{a logical indicating whether outer snps should be removed. Default is T.}

\item{snp:}{number of snps to include per simulation}

\item{form:}{output data frame in "wide" or "tall" form. This is an optional argument with default="wide".}

\item{fun:}{option to apply a function over the ss across a simulation. Default is "none". The haplotype statistics (h_stats) don't get transformed. Options include "norm"}

\item{LD_downsample:}{logical option to downsample genome matrix for the purposes of computing LD statistics. Default is FALSE.}

\item{ds_prop:}{Used with LD_downsample. Proportion of columns to downsample without replacement from the genome matrix.}

\item{ds_seed:}{Optional. Numeric seed for the LD_downsample.}

\item{missing_rate}{: Probability of elements that are randomly sampled to become NA for each row in G.}

\item{trans_prop:}{Proportion of columns (i.e. sites) that are chosen to be transition sites.
Default is 0.776.}

\item{dmg_rate:}{Probability of a element in a transition column changing from 0 to 1, or 1 to 0. 
Default is 0.05.}

\item{seed:}{a random seed for aging DNA.}
}
\value{
A one row dataframe summary stats and information for the input simulation object.
}
\description{
Takes in a single simulation object, ages the DNA and outputs summary statistics with the 
sum_stats function. The aging process consists of simulating missing information, pseduo-
haplotypes, deamination/DNA damage and ascertainment bias. See the age_DNA, pseudo_hap and
ascertain_bias function documentation for more information. See sum_stats function 
documentation to see how summary statistics are computed.
}
\examples{
sum_stats(sim_obj)
}
