% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{any_sug}
\alias{any_sug}
\title{any_sug function}
\usage{
any_sug(x)
}
\arguments{
\item{x:}{LogicalVector.}
}
\value{
A boolean.
}
\description{
A wrapper for the sugar function any. It takes a LogicalVector and returns T if any of the elements are T.
Source: https://gallery.rcpp.org/articles/sugar-any/
}
\examples{
 x <- c(3, 9, 0, 2, 7, -1, 6)
any_sug(x>9)
}
