% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ascertain_bias.R
\name{ascertain_bias}
\alias{ascertain_bias}
\title{ascertain_bias function}
\usage{
ascertain_bias(G, index)
}
\arguments{
\item{G}{: A binary genome matrix.}

\item{index}{: A vector of indices indicating the rows of G which represent the outgroup.}
}
\value{
A genome matrix consisting of the columns of G where the outgroup was hetereozygous.
The rows belonging to the outgroup are also excluded.
}
\description{
Function to simulate ascertainment bias in ancient DNA. Takes in a genome matrix G
where some rows represent sequences from an outgroup (usually 2). The function 
identifies all heterozygous sites in the outgroup, then subsets the G at these 
columns. The outgroup is excluded in the output matrix.
}
\examples{
G <-matrix(sample(0:1, size = 40, replace = TRUE), nc = 8)
ascertain_bias(G, c(7,8))
}
