% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clus_hap.R
\name{clus_hap}
\alias{clus_hap}
\title{clus_hap function}
\usage{
clus_hap(G, max_clus)
}
\arguments{
\item{max_clus}{: Integer for the maximum number of clusters to consider. If the number of 
unique rows is smaller than max_clus, the number of unique rows will be used instead. Must 
be at least 3.}

\item{G:}{Binary genome matrix of 0's and 1's. Each column is a SNP, each row is an individual.}
}
\value{
A Numeric vector indicating the clusters to which each row belongs
}
\description{
Takes a genome matrix and clusters the rows. The best number of clusters is chosen
by having the highest silhouette value. Returns a clustering vector. The i^th element
of the cluster vector is the cluster to which row i has been assigned.
}
\examples{
G = matrix(sample(0:1, size =25 , replace = TRUE), nc = 5)
clus_hap(G,3)
}
