% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clus_hstats.R
\name{clus_hstats}
\alias{clus_hstats}
\title{clus_hstats function}
\usage{
clus_hstats(clus_vec)
}
\arguments{
\item{clus_vec:}{A numeric vector indicating to which cluster, each row in the original 
genome matrix was assigned. The clusters are labelled from 1,2,3.... up to the some number 
of clusters.}
}
\value{
A numeric vector of haplotype statistics (h1,h2,h12,h123).
}
\description{
Takes in a cluster vector of a genome matrix and computes the haplotype statistics.
}
\examples{
G = matrix(sample(0:1, size =25 , replace = TRUE), nc = 5)
x = clus_hap(G,3)
clus_hstats(x)
}
