% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_sfs.R
\name{compute_sfs}
\alias{compute_sfs}
\title{ancient_sum_stats function}
\usage{
compute_sfs(
  sim,
  nwins = 1,
  missing_rate,
  trans_prop = 0.776,
  dmg_rate = 0.05,
  ascertain_indices,
  seed = NA,
  impute_method,
  trim_sim = F,
  snp = NA
)
}
\arguments{
\item{missing_rate}{: Probability of elements that are randomly sampled to become NA for each row in G.}

\item{sim:}{a simulation object}

\item{nwins:}{number of subwindows to split each genome matrix within the simulation. Default is 1.}

\item{trim_sim:}{a logical indicating whether outer snps should be removed. Default is F.}

\item{snp:}{number of snps to include per simulation}

\item{trans_prop:}{Proportion of columns (i.e. sites) that are chosen to be transition sites.
Default is 0.776.}

\item{dmg_rate:}{Probability of a element in a transition column changing from 0 to 1, or 1 to 0. 
Default is 0.05.}

\item{ascertain_indices:}{A list of 2-dimensional vectors. Each vector contains the indices of 2 
rows in G for doing the ascertainment bias.}

\item{impute_method:}{A string indicating the imputation method for missingness. Options are
"random" and "zero." See documentation on random_impute and zero_impute for more information.}

\item{seed:}{Optional. A random seed for aging DNA.}
}
\value{
A list of NumericVectors containing the SFS for each window/block.
}
\description{
Takes in a single simulation object, ages the DNA and outputs the SFS for each window. 
See the documentation of ancient_sum_stats to see how DNA ageging is simulated.
}
\examples{
ancient_sum_stats(sim_obj)
}
