% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{count}
\alias{count}
\title{count function
counts the number of times a value appears in an IntegerVector}
\usage{
count(vec, target)
}
\arguments{
\item{vec:}{a NumericVector}

\item{target:}{input integer value to find}
}
\value{
scalar value of the number of times the target appears in vec
}
\description{
count function
counts the number of times a value appears in an IntegerVector
}
\examples{
count(c(1,2,3,4),3)
}
