% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discoal_sim.R
\name{discoal_sim}
\alias{discoal_sim}
\title{discoal_sim function}
\usage{
discoal_sim(
  mu,
  recomb_rate,
  Ne,
  genome_length,
  samplesize,
  s = 0,
  discoal_path,
  fix_generation = NA,
  seed,
  sweep,
  start_freq = NA,
  popsize_changes = NULL
)
}
\arguments{
\item{mu}{mutation rate per base per generation. Warning! Check this again later in the discoal manual!}

\item{recomb_rate}{The recombination rate per base per generation}

\item{Ne}{The effective population size}

\item{genome_length}{The number of bases to simulate for each sample.}

\item{samplesize}{Number of samples to take from the population}

\item{s}{selection coefficient for the selected mutation. Default is 0. Note that for neutral simulations s must be 0.}

\item{discoal_path}{path to your discoal program}

\item{fix_generation}{number of generations ago when the selected mutation was fixed. Default value is 0.}

\item{seed}{vector of 2 numbers used for the simulations}

\item{sweep}{the kind of selective sweep. Options are "hard", "soft", "neutral" and "neutral_fixation".}

\item{start_freq}{Used for soft sweeps only. The mutation spreads via drift (neutral) and becomes selected only once it has reached the starting frequency.}

\item{popsize_changes}{A tibble with a size and a time column. The size is a multiplier for the current population size. The correponding time is the time
of the change in generations.}
}
\value{
an object of class sim_obj. Here are the features. cmd is the command. Seeds: the seeds used in the discoal simulation.
num_seg: number of segregating sites in the sampled population. pos: vector of the positions of every seg site (infinite sites model)
sweep: the kind of selective sweep modelled. s: the selection coefficient
}
\description{
discoal_sim function
}
\examples{
sim<-discoal_sim(mu=mu,recomb_rate=recomb_rate,Ne=Ne,genome_length=genome_length,samplesize=samplesize,s=s,discoal_path=discoal_path,fix_generation=fix,sweep=sweep)

}
