% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{fill_row}
\alias{fill_row}
\title{fill_row function
takes a NumericVector and adds it as a new row on the bottom of a matrix}
\usage{
fill_row(A, x)
}
\arguments{
\item{A:}{A NumericMatrix.}

\item{x:}{A NumericVector.}
}
\description{
fill_row function
takes a NumericVector and adds it as a new row on the bottom of a matrix
}
