% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{find_index}
\alias{find_index}
\title{find_index function}
\usage{
find_index(x, target)
}
\arguments{
\item{x:}{A NumericVector with the elements are arranged in ascending order.}

\item{target:}{a double}
}
\value{
An integer for the index of the closest element in NumericVector x
}
\description{
Searches a NumericVector for the element that is closest to a target value. 
Returns the corresponding index.
}
\examples{
x<-runif(0,1,n=5) \%>\% sort()
target=0.2
}
