% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{first_over}
\alias{first_over}
\title{first_over function
Takes in a sorted NumericVector and a target double. Find the first element
in the vector which is larger than the target. Returns the index.}
\usage{
first_over(x, target)
}
\arguments{
\item{x:}{A sorted NumericVector}

\item{target:}{A double.}
}
\value{
An integer for the index of the first element in x that is larger than the target.
}
\description{
first_over function
Takes in a sorted NumericVector and a target double. Find the first element
in the vector which is larger than the target. Returns the index.
}
\examples{
x<-runif(0,1,n=5) \%>\% sort() 
target=0.2 
find_index(x,target)
}
