% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{G_flip}
\alias{G_flip}
\title{G_flip function}
\usage{
G_flip(G)
}
\arguments{
\item{:}{G: Binary genome matrix of 0's and 1's. Each column is a SNP, each row is an individual.}
}
\value{
: A NumericMatrix.
}
\description{
Denoises a genome matrix by flipping an element in each column. For each column, 
we flip the element which results in the smallest number of pairwise differences
in the whole matrix. If there is a tie, we flip the element closest to the top.
}
