% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{Gcol_denoise}
\alias{Gcol_denoise}
\title{Gcol_denoise function}
\usage{
Gcol_denoise(G, col)
}
\arguments{
\item{:}{col: A column index. 0 indexing is used.}
}
\value{
: A NumericMatrix like G, with one element flipped.
}
\description{
Takes a matrix and a column index. Flips the element in that column which leads
to the greatest reduction in the number of pairwise differences. When there is 
a tie, the first element among the tied elements is changed.
}
