% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{Gcol_flip}
\alias{Gcol_flip}
\title{Gcol_flip function}
\usage{
Gcol_flip(G, col)
}
\arguments{
\item{G:}{Binary genome matrix of 0's and 1's. Each column is a SNP, each row is an individual.}

\item{col:}{A column index. 0 indexing is used.}
}
\value{
A NumericVector where element i is the number of pairwise differences in G if 
element i in the selected column was flipped.
}
\description{
Takes binary genome matrix and a column index. Returns the number of pairwise
differences that would occur if the individual elements of the column were 
flipped.
}
