% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_df.R
\name{generate_df}
\alias{generate_df}
\title{generate_df function}
\usage{
generate_df(
  sim_list,
  nwins,
  split_type = "base",
  snp,
  form = "wide",
  fun = "none",
  LD_downsample = F,
  ds_prop = NA,
  ds_seed = NA
)
}
\arguments{
\item{sim_list:}{list of simulation objects}

\item{nwins:}{number of subwindows desired to split each genome matrix per simulation}

\item{split_type:}{Method of splitting the genome matrix. Valid options are "base" and "mut". Default "base".}

\item{snp:}{number of snps to include per simulation}

\item{form:}{Option for dataframe to be in "tall" or "wide" form.}

\item{fun:}{function to apply over summary statistics (haplotype stats excluded). See sum_stats documentation for details.}

\item{LD_downsample:}{logical option to downsample genome matrix for the purposes of computing LD statistics. Default is FALSE.}

\item{ds_prop:}{Used with LD_downsample. Proportion of columns to downsample without replacement from the genome matrix.}

\item{ds_seed:}{Optional. Numeric seed for the LD_downsample.}
}
\value{
dataframe containing summary statistics of all the subwindows across all the simulations.
}
\description{
Takes a list of simulation objects and produces a data frame. The genome matrix of each simulation is broken into a series of non-overlapping subwindows and SS are computed. Each subwindow is a row on the final dataframe.
}
\examples{
generate_df(sim_list,10)
}
