% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{h_stats}
\alias{h_stats}
\title{h_stats function}
\usage{
h_stats(G)
}
\arguments{
\item{G:}{A binary genome matrix consisting of 1's and 0's. Each column is a SNP. Each row is a sampled individual.}
}
\value{
A numeric vector of h stats. (h1,h2,h12)
}
\description{
Computes the h1,h2,h12 statistics. For more information see https://doi.org/10.1371/journal.pgen.1005004.
}
\examples{
h_stats(G)
}
