% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{have_na}
\alias{have_na}
\title{have_na function}
\usage{
have_na(M)
}
\arguments{
\item{M:}{A NumericMatrix}
}
\value{
A boolean indicating whether the input matrix has any NAs.
}
\description{
Checks if a NumericMatrix has any NA values. Returns T if it does, F otherwise.
}
\examples{
A = matrix(1:4, nrow = 2)
have_na(A)
}
