% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_genome_matrix.R
\name{is_genome_matrix}
\alias{is_genome_matrix}
\title{is_genome_matrix function}
\usage{
is_genome_matrix(x)
}
\arguments{
\item{x:}{an R object}
}
\value{
a logical indicating whether the input is a valid genome matrix
}
\description{
Checks if an R object is a valid genome matrix. A genome matrix is a binary 
numeric matrix consisting of only 0's and 1's. Each column must have at least
one entry which is 1.
}
\examples{
seq <-matrix(sample(0:1, size =16 , replace = TRUE), nc = 4)
is_genome_matrix(seq)
}
