% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{is_sorted}
\alias{is_sorted}
\title{is_sorted function}
\usage{
is_sorted(x, ascend)
}
\arguments{
\item{x:}{NumericVector}

\item{ascend:}{A boolean. If TRUE, checks if the elements of x are in ascending order.
If FALSE, checks if they are in descending order.}
}
\value{
A boolean indicating whether the vector is sorted.
}
\description{
Checks if a NumericVector is sorted numerically.
}
\examples{
x = c(1,3,5,7)
is_sorted(x, ascend = T)
}
