% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LD_w.R
\name{LD_w}
\alias{LD_w}
\title{LD_w function}
\usage{
LD_w(r, i)
}
\arguments{
\item{r}{: correlation matrix showing the correlations between each SNP 
position, squared.}

\item{i}{:index for a column in r. i must be bigger than 1 and less than the number
of columns in r-1.}
}
\value{
a numeric value for the omega statistic between at SNP column i.
}
\description{
Computes w for some SNP position i, in a genome matrix.
}
\examples{
  mat <- matrix(runif(64), 8, 8)
mat[upper.tri(mat)==F] <- 0
LD_w(mat,2)
}
