% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/majority_flip.R
\name{majority_flip}
\alias{majority_flip}
\title{majority_flip function}
\usage{
majority_flip(G, seed)
}
\arguments{
\item{G:}{Binary genome matrix of 0's and 1's. Each column is a SNP, each row is an individual.}

\item{seed:}{A random seed.}
}
\value{
A NumericMatrix where one element has been randomly switched in each column.
}
\description{
Takes a binary NumericMatrix of 0's and 1's. For each column, one entry from the
minority group is randomly selected to be switched to the majority.
}
\examples{
G = matrix(sample(0:1, size =20 , replace = TRUE), nc = 5)
majority_flip(G)
}
