% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix2genotype.R
\name{matrix2genotype}
\alias{matrix2genotype}
\title{matrix2genotype function}
\usage{
matrix2genotype(G)
}
\arguments{
\item{G:}{A binary genome matrix. Columns are mutations, rows are the sampled chromosomes. 
G must have an even number of rows.}
}
\value{
A dataframe of genotype objects.
}
\description{
Converts a binary genome matrix into a dataframe of genotype S3 objects as defined
in the genetics package in R. The first pair of rows from the top will be the first 
genotype.The second pair of rows will be the second genotype and so on. Hence, the 
input matrix must have an even number of rows.
}
\examples{
insert later

}
