% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{matrix_subset}
\alias{matrix_subset}
\title{matrix_subset function}
\usage{
matrix_subset(G, y)
}
\arguments{
\item{G:}{A NumericMatrix}

\item{y:}{An IntegerVector of indices. Uses 1 indexing to suit R.}
}
\value{
A NumericMatrix with subset columns of G using the indices in y.
}
\description{
Subsets selected columns from a NumericMatrix.
}
