% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/no_scientific.R
\name{no_scientific}
\alias{no_scientific}
\title{no_scientific function}
\usage{
no_scientific(x)
}
\arguments{
\item{x}{some numeric scalar}
}
\value{
the number x as a character vector without scientific notation
}
\description{
Changes a number away from scientific notation. Discoal doesn't like scientific notation.
}
\examples{
change 1e6 to 1000000
num<-no_scientific(1e6)

}
