% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/norm_vec.R
\name{norm_vec}
\alias{norm_vec}
\title{norm_vec function}
\usage{
norm_vec(vec)
}
\arguments{
\item{vec:}{a numeric vector}
}
\value{
a normalised numeric vector
}
\description{
Normalizes a numeric vector by taking away the mean and dividing by the standard deviation, for each element.
}
\examples{
x<-rep(1:10)
norm_vec(x)
}
