% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{pairwise_diff}
\alias{pairwise_diff}
\title{pairwise_diff}
\usage{
pairwise_diff(G)
}
\arguments{
\item{G:}{Binary genome matrix of 0's and 1's. Each column is a SNP, each row is an individual.}
}
\value{
scalar value of the number of pairwise differences
}
\description{
Computes the number of pairwise differences between all the rows of a 
NumericMatrix.
}
\examples{
G = matrix(sample(0:1, size =5*5 , replace = TRUE), nc = 5)
}
