\name{popgen.tools-package}
\alias{popgen.tools-package}
\alias{popgen.tools}
\docType{package}
\title{
  Popgen Tools Package
}
\description{
  This package uses the discoal package to generate coalescent simulations   and save them as sim objects. Functions are available to compute summary   statistics based on the infinite sites model. In discoal, each column in   genome matrix is a SNP and each row is an individual. Hence, the summary   statistic functions are written for genome matrices of this form. 
}
\details{
  This section should provide a more detailed overview of how to use the
  package, including the most important functions.
}
\author{
Anthony Kwong

Maintainer: Anthony Kwong <a1708050@adelaide.edu.au>
}
\references{
  This optional section can contain literature or other references for
  background information.
}
\keyword{ package }
\seealso{
  Optional links to other man pages
}
\examples{
  \dontrun{
     ## Optional simple examples of the most important functions
     ## These can be in \dontrun{} and \donttest{} blocks.   
  }
}
