% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{present_row}
\alias{present_row}
\title{present_row function}
\usage{
present_row(A, x)
}
\arguments{
\item{A:}{NumericMatrix}

\item{x:}{NumericVector}
}
\value{
integer reocrding how many times the vector is present as a row in the matrix
}
\description{
checks if a vector is present as a row in the input matrix
}
