% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pseudo_hap.R
\name{pseudo_hap}
\alias{pseudo_hap}
\title{pseudo_hap function}
\usage{
pseudo_hap(G, seed = NA)
}
\arguments{
\item{G}{: A binary genome matrix: The number of rows must be even.}

\item{seed:}{A numeric random seed.}
}
\value{
: A smaller genome matrix consisting of the pseudo-haplotypes of G. Output has
half the number of rows as G.
}
\description{
Simulates pseudo haplotypes for a genome matrix. The matrix is broken into pairs of 
rows. For each row pair, we randomly sample an element for each column to make
new pseudo haplotypes. Function outputs a matrix of pseudo haplotypes with half
the number of rows as the original genome matrix.
}
\examples{

}
