% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{random_impute}
\alias{random_impute}
\title{random_impute function}
\usage{
random_impute(G)
}
\arguments{
\item{G:}{A NumericMatrix with NAs.}
}
\value{
A NumericMatrix with imputed values in place of the NAs.
}
\description{
Imputes NAs in a genome matrix. For each NA, we randomly sample from the non-NA
elements of that column. Each element is equally likely to be selected. NA becomes
the randomly selected element.
}
\examples{
G <-matrix(sample(0:1, size = 16, replace = TRUE), nc = 4)
G[1,1] = NA
random_impute(G)
}
