% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{row_count}
\alias{row_count}
\title{row_count function}
\usage{
row_count(A, x)
}
\arguments{
\item{A:}{NumericMatrix}

\item{x:}{NumericVector}
}
\value{
integer recording how many times the vector is present as a row in the matrix
}
\description{
Counts how many times a vector is present as a row in a NumericMatrix.
}
