% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{row_freq}
\alias{row_freq}
\title{row_freq function}
\usage{
row_freq(A)
}
\arguments{
\item{NumericMatrix}{A}
}
\value{
A NumericVector containing the frequency of each unique row in A. The order of this vector is the order by which the unique rows appear in A.
}
\description{
Takes a NumericMatrix and returns the frequency of each unique row
}
\examples{
row_freq(A)
}
