% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/row_replicate.R
\name{row_replicate}
\alias{row_replicate}
\title{row_replicate function}
\usage{
row_replicate(M, row, n)
}
\arguments{
\item{M:}{A numeric matrix}

\item{row:}{the index of the row  to replicate}

\item{n:}{number of times to replicate row}
}
\value{
A numeric matrix with selected row appended at the bottom
}
\description{
Replicates a selected row in a matrix n times and appends it to the bottom.
}
\examples{
row_replicate(M,row=2,n=3)
}
