% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_vec.R
\name{sample_vec}
\alias{sample_vec}
\title{sample_vec function}
\usage{
sample_vec(x)
}
\arguments{
\item{x:}{A vector.}
}
\value{
An random element selected from the input vector.
}
\description{
Randomly samples an element from a vector. If the vector contains only one element, that
element is returned.
}
\examples{
x <- c(1,3,4,5)
sample_vec(x)
}
