% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_obj.R
\name{sim_obj}
\alias{sim_obj}
\title{sim_obj function}
\usage{
sim_obj(cmd, seeds, segsites, positions, genome_matrix, sweep,
  select_coeff, fix_generation)
}
\arguments{
\item{cmd}{the command used to generate the simulation}

\item{seeds}{}

\item{segsites}{}

\item{positions}{}

\item{genome_matrix}{}

\item{sweep}{}

\item{select_coeff}{}

\item{fix_generation}{}
}
\value{
an object called sim_obj
}
\description{
Generate a simulation object to store the data produced by discoal.
}
