% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snp_count.R
\name{snp_count}
\alias{snp_count}
\title{snp_count function}
\usage{
snp_count(sim_list)
}
\arguments{
\item{sim_list:}{a list of simulation objects}
}
\value{
A dataframe with 2 columns (sweep_type, selection_coeff, SNP_count)
}
\description{
snp_count function
}
\examples{
snp_count(sim_list)
}
