% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/string_labels.R
\name{string_labels}
\alias{string_labels}
\title{string_labels function}
\usage{
string_labels(f, it)
}
\arguments{
\item{f:}{a character string}

\item{it:}{number of iterations}
}
\value{
a vector of character strings of f_1,f_2,....,f_n
}
\description{
Generates a string vector f_1,f_2,....,f_n
}
\examples{
string_labels("Echo",5)
}
