% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sum_stats.R
\name{sum_stats}
\alias{sum_stats}
\title{sum_stats function}
\usage{
sum_stats(
  sim,
  nwins = 1,
  split_type = "base",
  ID,
  trim_sim = T,
  snp = NA,
  form = "wide",
  fun = "none",
  LD_downsample = F,
  ds_prop = NA,
  ds_seed = NA
)
}
\arguments{
\item{sim:}{a simulation object}

\item{nwins:}{number of subwindows to split each genome matrix within the simulation. Default is 1.}

\item{split_type:}{Method of splitting the genome matrix. Valid options are "base" and "mut". Default is "base".}

\item{ID:}{a numeric ID value to group subwindows under the simulation it came from.}

\item{trim_sim:}{a logical indicating whether outer snps should be removed. Default is T.}

\item{snp:}{number of snps to include per simulation}

\item{form:}{output data frame in "wide" or "tall" form. This is an optional argument with default="wide".}

\item{fun:}{option to apply a function over the ss across a simulation. Default is "none". The haplotype statistics (h_stats) don't get transformed. Options include "norm"}

\item{LD_downsample:}{logical option to downsample genome matrix for the purposes of computing LD statistics. Default is FALSE.}

\item{ds_prop:}{Used with LD_downsample. Proportion of columns to downsample without replacement from the genome matrix.}

\item{ds_seed:}{Optional. Numeric seed for the LD_downsample.}
}
\value{
list of summary stats for the input genome matrix
}
\description{
Turns a simulation object into a data frame. The genome matrix is partitioned into subwindows and summary statistics are computed. Each subwindow is a row on the output tibble.
}
\examples{
sum_stats(win_list)
This is meant to be a hidden function. Hide in final version. 
}
