% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{theta_h}
\alias{theta_h}
\title{theta_h function}
\usage{
theta_h(G)
}
\arguments{
\item{G:}{Binary genome matrix with 0's and 1's. Each column is a SNP, each row is an individual.}
}
\value{
scalar value of theta_h for that sampled population
}
\description{
Computes theta_h for a genome matrix
}
