% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{theta_t}
\alias{theta_t}
\title{theta_t function}
\usage{
theta_t(G)
}
\arguments{
\item{G:}{G: Binary genome matrix of 0's and 1's. Each column is a SNP, each row is an individual.}
}
\value{
scalar value of theta_t
}
\description{
Computes theta_t, the number of pairwise differences normalised by the number of pairs.
}
\examples{
theta_t(G)
}
