% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{theta_w}
\alias{theta_w}
\title{theta_w function}
\usage{
theta_w(G)
}
\arguments{
\item{G:}{Binary genome matrix of 0's and 1's. Each column is a SNP, each row is an individual.}
}
\value{
scalar value of theta_w
}
\description{
Computes the average number of pairwise differences, assuming neutrality.
}
\examples{
theta_w(G)
}
