% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{unique_rows}
\alias{unique_rows}
\title{unique_rows function}
\usage{
unique_rows(A)
}
\arguments{
\item{A:}{A general matrix of real values.}
}
\value{
A NumericVector. i'th element is the frequency of the i'th unique row.
}
\description{
Takes a NumericMatrix and returns the frequency of all the unique rows as a NumericVectior.
}
\examples{
unique_rows(A)
}
