% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{var_taj}
\alias{var_taj}
\title{var_taj function}
\usage{
var_taj(G)
}
\arguments{
\item{G:}{Binary genome matrix of 0's and 1's. Each column is a SNP, each row is an individual.}
}
\value{
scalar value of the variance term in tajima's D.
}
\description{
Computes the variance of the nominator term in tajima's D.
}
\examples{
var_taj(G)
}
