% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{vec_equal}
\alias{vec_equal}
\title{vec_equal function for cpp}
\usage{
vec_equal(x, y)
}
\arguments{
\item{x, y:}{two numeric vectors of the same dimensions}
}
\value{
boolean, true if the two vectors are the same. False otherwise.
}
\description{
vec_equal function for cpp
}
\examples{
vec_equal(x,y)
}
