% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{vec_sort}
\alias{vec_sort}
\title{vec_sort function}
\usage{
vec_sort(x)
}
\arguments{
\item{x:}{NumericVector}
}
\value{
NumericVector with elements of x sorted in descending order
}
\description{
Sorts the elements of a NumericVector in descending order
}
\examples{
x<-c(1,8,2,2,3)
vec_sort(x)
}
