% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{vec_split}
\alias{vec_split}
\title{vec_split function}
\usage{
vec_split(x, n)
}
\arguments{
\item{x:}{vector to split}

\item{n:}{number of blocks to split a vector into.}
}
\value{
a list of vectors
}
\description{
Breaks a vector x into n equal sized blocks. If the dimension of x is not divisible by n, we put the remaining elements into the last block.
}
\examples{
vec_split(seq(10),3)
}
