% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{vector_trim}
\alias{vector_trim}
\title{vector_trim function}
\usage{
vector_trim(x, cen, k)
}
\arguments{
\item{x:}{NumericVector to trim}

\item{cen:}{Integer index for central element}

\item{k:}{Number of elements to retain from both sides of the center.}
}
\value{
A NumericVector with outer elements removed. Up to k elements can be 
retained from both sides of the center.
}
\description{
Trims outer elements of a vector based on some provided central index.
}
