% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{w_max}
\alias{w_max}
\title{w_max function}
\usage{
w_max(x)
}
\arguments{
\item{G:}{Binary genome matrix consisting of 1's and 0's.}
}
\value{
scalar value of w_max for the genome matrix
}
\description{
Computes w_max for a binary, genome matrix. Each column is a SNP, each row is a sample.
}
\examples{
w_max(G)
}
