% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{window_trim}
\alias{window_trim}
\title{window_trim function}
\usage{
window_trim(G, cen, k)
}
\arguments{
\item{G:}{NumericMatrix}

\item{cen:}{An integer designating a column index. This will be the center column of the ouput matrix.}

\item{k:}{An integer for the number of columns to include from either side of the center column.}
}
\value{
A NumericMatrix with the outer columns of G removed. Up to k/2 columns to either side of the center column from the original matrix is kept.
}
\description{
Trims the outer columns of a matrix.
}
\examples{
window_trim(G,1,5)
}
