% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{winsplit_base}
\alias{winsplit_base}
\title{winsplit_base function}
\usage{
winsplit_base(G, pos, n)
}
\arguments{
\item{G:}{A NumericMatrix designating a binary genome matrix consisting of 1's and 0's. Each column is a SNP. Each row is a sampled individual.}

\item{pos:}{A NumericVector consisting of values between 0,1. The ith element is the position of the ith SNP in G.}

\item{n:}{integer for the number of desired windows in the output list.}
}
\value{
A NumericMatrix list of the windows
}
\description{
Split a genome matrix into equally sized windows as determined by
the number of bases. The windows can differ in the number of columns
because SNPs are not always uniform across a given region.
}
\examples{
seq <-matrix(sample(0:1, size = 25, replace = TRUE), nc = 5)
pos<-runif(0,1,n=5) \%>\% sort()
winsplit_length(seq,pos,5)
}
