% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{winsplit_length}
\alias{winsplit_length}
\title{winsplit_length function}
\usage{
winsplit_length(G, pos, n)
}
\arguments{
\item{G:}{A NumericMatrix designating a binary genome matrix consisting of 1's and 0's. Each column is a SNP. Each row is a sampled individual.}

\item{pos:}{A NumericVector consisting of values between 0,1. The ith element is the position of the ith SNP in G.}

\item{n:}{integer for the number of desired windows in output list.}
}
\value{
A NumericMatrix list of the windows
}
\description{
Splits a genome matrix into equally sized windows as determined by the number of bases. 
The windows can differ in their number of columns when SNPs are not uniformly distributed
across the region.
}
\examples{
 
seq <-matrix(sample(0:1, size = 25, replace = TRUE), nc = 5)
pos<-runif(0,1,n=5) \%>\% sort()
winsplit_length(seq,pos)
}
