% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zero_impute.R
\name{zero_impute}
\alias{zero_impute}
\title{zero_impute function}
\usage{
zero_impute(G)
}
\arguments{
\item{G:}{A NumericMatrix with NAs.}
}
\value{
A new matrix where all the NA elements of G have been replaced by 0.
}
\description{
Changes all NAs in a matrix into 0. A conservative approach as it assumes that 
there is no derived allele present where there is missing information.
}
\examples{
G <- rbind(c(1, 0, 1, NA, 0), c(1, NA, NA, 0, 1), c(0, 0, 1, 0, NA))
 zero_impute(G)
}
