% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/age_DNA.R
\name{age_DNA}
\alias{age_DNA}
\title{age_DNA function}
\usage{
age_DNA(G, missing_rate, trans_prop = 0.776, dmg_rate = 0.05, seed = NA)
}
\arguments{
\item{G}{: A binary, numeric genome matrix.}

\item{missing_rate}{: Probability of elements that are randomly sampled to become NA for each row in G.}

\item{trans_prop:}{Proportion of columns (i.e. sites) that are chosen to be transition sites.
Default is 0.776.}

\item{dmg_rate:}{Probability of a element in a transition column changing from 0 to 1, or 1 to 0. 
Default is 0.05.}

\item{seed:}{A numeric random seed. Optional.}
}
\value{

}
\description{
A function to simulating DNA aging on discoal simulations. To simulate missingness, for each row 
a percentage of elements are randomly sampled and become missing. To simulate deamination
a percentage of columns are randomly sampled according to trans_prop, to become transition sites. 
For each transition site we flip a fair coin. If heads we change 0's to 1's, with probability as 
specified by the dmg_rate. If tails, we change 1's to 0's, with probability as 
specified by the dmg_rate.
}
\details{
This method is adapted from doi: 10.1534/genetics.112.139949
}
\examples{

}
