% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ancient_sum_stats.R
\name{ancient_sum_stats}
\alias{ancient_sum_stats}
\title{ancient_sum_stats function}
\usage{
ancient_sum_stats(
  sim,
  nwins = 1,
  split_type = "base",
  ID,
  trim_sim = F,
  snp = NA,
  missing_rate,
  trans_prop = 0.776,
  dmg_rate = 0.05,
  ascertain_indices,
  seed = NA,
  impute_method
)
}
\arguments{
\item{missing_rate}{: Probability of elements that are randomly sampled to become NA for each row in G.}

\item{sim:}{a simulation object}

\item{nwins:}{number of subwindows to split each genome matrix within the simulation. Default is 1.}

\item{split_type:}{Method of splitting the genome matrix. Valid options are "base" and "mut". Default is "base".}

\item{ID:}{a numeric ID value to group subwindows under the simulation it came from.}

\item{trim_sim:}{a logical indicating whether outer snps should be removed. Default is F.}

\item{snp:}{number of snps to include per simulation}

\item{trans_prop:}{Proportion of columns (i.e. sites) that are chosen to be transition sites.
Default is 0.776.}

\item{dmg_rate:}{Probability of a element in a transition column changing from 0 to 1, or 1 to 0. 
Default is 0.05.}

\item{ascertain_indices:}{A list of 2-dimensional vectors. Each vector contains the indices of 2 
rows in G for doing the ascertainment bias.}

\item{impute_method:}{A string indicating the imputation method for missingness. Options are
"random" and "zero." See documentation on random_impute and zero_impute for more information.}

\item{seed:}{Optional. A random seed for aging DNA.}
}
\value{
A one row dataframe summary stats and information for the input simulation object.
}
\description{
Takes in a single simulation object, ages the DNA and outputs summary statistics with the 
sum_stats function. The aging process consists of simulating missing information, pseduo-
haplotypes, deamination/DNA damage and ascertainment bias. Multiple pairs of individuals 
can be used for ascertainment bias. In that case, we will subset the columns of G which are
heterozygous in any of the pairs of individuals. See the age_DNA, pseudo_hap and
ascertain_bias function documentation for more information. See sum_stats function 
documentation to see how summary statistics are computed.
}
\examples{
ancient_sum_stats(sim_obj)
}
