% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discoal_sim.R
\name{discoal_sim}
\alias{discoal_sim}
\title{discoal_sim function}
\usage{
discoal_sim(
  mu,
  recomb_rate,
  Ne,
  genome_length,
  samplesize,
  s = 0,
  discoal_path,
  fix_time = NA,
  seed,
  sweep,
  start_freq = NA,
  popsize_changes = NULL,
  demes = NA,
  sample_dist = NA,
  deme_join = NULL
)
}
\arguments{
\item{mu}{mutation rate per base per generation. Warning! Check this again later in the discoal manual!}

\item{recomb_rate}{The recombination rate per base per generation}

\item{Ne}{The effective population size}

\item{genome_length}{The number of bases to simulate for each sample.}

\item{samplesize}{Total number of samples to take from the population}

\item{s}{selection coefficient for the selected mutation. Default is 0. Note that for neutral simulations s must be 0.}

\item{discoal_path}{path to your discoal program}

\item{fix_time}{number of generations ago when the selected mutation was fixed.}

\item{seed}{vector of 2 numbers used for the simulations}

\item{sweep}{the kind of selective sweep. Options are "hard", "soft", "neutral" and "neutral_fixation".}

\item{start_freq}{Used for soft sweeps only. The mutation spreads via drift (neutral) and becomes selected only once it has reached the starting frequency.}

\item{popsize_changes}{A tibble with a size and a time (measured generations) column. The size is a multiplier for the current population size. The correponding time is the time
of the change in generations. Current version supports 2 changes per simulation to model bottlenecks. Discoal normally uses time in units of 4Ne but this function does the conversion.}

\item{demes}{Optional.A numeric integer indicating the number of population demes to simulate. Only one deme is simulated by default.}

\item{sample_dist}{Optional. Only usable when simulating multiple demes. A numeric integer vector indicating the number of samples to make from each deme. 
The sum must be equal to samplesize.}

\item{deme_join}{Optional. A tibble with time, pop1, pop2 columns, indicating the time (measured in generations) to join 2 particular demes. Time is numeric. Demes are numeric
integers indicating the index of the deme. Note that discoal uses 0 indexing for the demes. Used to join demes if they are present. Discoal normally uses time in units of 4Ne but 
this function does the conversion so that time is entered in generations.}
}
\value{
an object of class sim_obj. Here are the features. cmd is the command. Seeds: the seeds used in the discoal simulation.
num_seg: number of segregating sites in the sampled population. pos: vector of the positions of every seg site (infinite sites model)
sweep: the kind of selective sweep modelled. s: the selection coefficient
}
\description{
discoal_sim function
}
\examples{
sim<-discoal_sim(mu=mu,recomb_rate=recomb_rate,Ne=Ne,genome_length=genome_length,samplesize=samplesize,s=s,discoal_path=discoal_path,fix_time=fix,sweep=sweep)

}
