% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{fwh}
\alias{fwh}
\title{fwh function
Computes Fay and Wu's H for a genome matrix.}
\usage{
fwh(x)
}
\arguments{
\item{G:}{Binary genome matrix with 0's and 1's. Each column is a SNP, each row is an individual.}
}
\value{
scalar value of Fay and Wu's H for that sampled population.
}
\description{
fwh function
Computes Fay and Wu's H for a genome matrix.
}
