% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_df.R
\name{generate_df}
\alias{generate_df}
\title{generate_df function}
\usage{
generate_df(sim_list, win_split)
}
\arguments{
\item{sim_list}{list of simulation objects}

\item{win_split}{number of subwindows desired to split each genome matrix per simulation}
}
\value{
dataframe containing summary statistics of all the subwindows across all the simulations.
}
\description{
Takes a list of simulation objects and produces a data frame. The genome matrix of each simulation is broken into a series of non-overlapping subwindows and SS are computed. Each subwindow is a row on the final dataframe.
}
