% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_obj.R
\name{sim_obj}
\alias{sim_obj}
\title{sim_obj function}
\usage{
sim_obj(
  cmd,
  seeds,
  segsites,
  positions,
  genome_matrix,
  sweep,
  select_coeff,
  fix_time = 0,
  bottle_time1 = 0,
  bottle_size1 = 1,
  bottle_time2 = 0,
  bottle_size2 = 1
)
}
\arguments{
\item{cmd}{The command used to generate the simulation}

\item{seeds}{The two random seeds used in the discoal simulation}

\item{segsites}{Number of segregating sites found in sampled population}

\item{positions}{A vector with elements between 0,1. Designates the position of each seg site.}

\item{genome_matrix}{Rows are samples. Columns are mutations/segregating sites.}

\item{sweep}{Type of sweep simulated.}

\item{select_coeff}{S value for the strength of the sweep.}

\item{fix_time}{The time of fixation of the fixed mutation. Not needed for neutral simulations.
Default is 0, indicating fixation at the time of sampling.}

\item{bottle_time1}{Time (in generations) of the most recent popsize change.}

\item{bottle_size1}{Multiplier for the sampled/reference population size. 
Refers to the most recent popsize change, looking back from the present.}

\item{bottle_time2}{Time (in generations) of the second most recent popsize change.}

\item{bottle_size2}{Multiplier for the sampled/reference population size. 
Refers to the second most recent popsize change, looking back from the present.}
}
\value{
an object called sim_obj
}
\description{
Generate a simulation object to store the data produced by discoal.
}
\examples{
This is meant to be a hidden function.
}
