% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{sub_win}
\alias{sub_win}
\title{sub_win function}
\usage{
sub_win(G, num_windows)
}
\arguments{
\item{G:}{A binary matrix of 0's and 1's.}

\item{num_win:}{Number of subwindows to break G into.}
}
\value{
A List of 2 elements. 1. A list of NumericMatrix for the windows. 
2. A NumericVector of column indices, indicating the boundary points for each window. 
We use the R setup where indices start at 1.
}
\description{
Breaks down a matrix into a series of equal sized, non-overlapping windows.
}
\examples{
sub_win(G,4)
}
