% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{sub_win}
\alias{sub_win}
\title{sub_win function
Breaks down a matrix into a series of equal sized, non-overlapping windows}
\usage{
sub_win(G, num_windows)
}
\arguments{
\item{G:}{A binary matrix of 0's and 1's.}

\item{num_win:}{Number of subwindows to break G into.}
}
\description{
sub_win function
Breaks down a matrix into a series of equal sized, non-overlapping windows
}
