% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sum_stats.R
\name{sum_stats}
\alias{sum_stats}
\title{sum_stats function}
\usage{
sum_stats(sim, win_split, ID, snp)
}
\arguments{
\item{sim:}{a simulation object}

\item{win_split:}{number of subwindows to split each genome matrix within the simulation.}

\item{ID:}{an ID value to group subwindows under the simulation it came from.}

\item{snp:}{number of snps to include per simulation}
}
\value{
list of summary stats for the input genome matrix
}
\description{
Turns a simulation object into a data frame. The genome matrix is partitioned into subwindows and summary statistics are computed. Each subwindow is a row on the output tibble.
}
\examples{
sum_stats(win_list)
This is meant to be a hidden function. Hide in final version. 
}
