% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sum_stats.R
\name{sum_stats}
\alias{sum_stats}
\title{sum_stats function}
\usage{
sum_stats(win_list)
}
\arguments{
\item{windows:}{A list of  binary genome matrices. Columns are SNPs and rows are samples.}
}
\value{
list of summary stats for the input genome matrix
}
\description{
Computes a variety of popgen summary stats on a genome matrix list.
}
\examples{
sum_stats(win_list)
This is meant to be a hidden function.
}
