% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{taj_D}
\alias{taj_D}
\title{taj_D function}
\usage{
taj_D(G)
}
\arguments{
\item{G:}{A binary matrix of 0's and 1's. Each column is a SNP and each row is a sampled individual.}
}
\value{
a scalar value of tajima's D for the sampled population.
}
\description{
taj_D function
}
\examples{
taj_D(G)
}
