% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{taj_D}
\alias{taj_D}
\title{taj_D function}
\usage{
taj_D(t_t, t_w, var_taj)
}
\arguments{
\item{t_t:}{theta_t for geneom matrix G. Use theta_t()}

\item{t_w:}{theta_w for genome matrix G. Use theta_w()}

\item{var_taj:}{Variance of tajima's D}
}
\value{
a scalar value of tajima's D for the sampled population.
}
\description{
Computes Tajima's D for a genome matrix.
}
\examples{
taj_D(theta_t(G),theta_w(G),var_taj(G))
}
