% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{unique_rows}
\alias{unique_rows}
\title{unique_rows function}
\usage{
unique_rows(A)
}
\arguments{
\item{A:}{A general matrix of real values.}
}
\value{
A NumericMatrix containing all the unique rows of A.
}
\description{
Takes a NumericMatrix and returns all the unique rows as a NumericVector. The rows appear in the order they were in A.
}
\examples{
unique_rows(A)
}
